Do not display age in email when birthday year is incoherent
birthday-pal relies on remote contact informations, and some informations can be inconsistent.

Sometimes, the year of birth is not specified or inconsitent (may be 1604 when year is null on some providers)

We consider that a contact can not be born before 1900 (120 years until now) in order to remove inconsistent age values in emails.