Remove Google and Oauth2 support
Interacting with Google APIS (HTTP and golang) is very painful and fragile.

It broke a few times:

- https://github.com/s-petit/birthday-pal/commit/1c897150115ac99da67579f4bd0904d923c80272

Now Oauth2 management via Google changed and not work anymore.

It is too painful to obtain authorization from Google in order to read people API (provide a home page, a youtube video, usage conditions...)

bpal wasn't designed to be widely deployed on the web and read dozens of personal data. It's just a small tool which read birthdays and send emails.